fix: link layer destroy calls callback
calling the callback passed to `_destroy` is not explicity mentioned as required in the nodejs docs, but it's necessary for error propagation.
this is why the tests listening for an error event from link layer would fail, because even though the link layer would be destroyed, it would not emit an error event.